refactor: se cambia tipo de dato

+++ back/PuntoVenta/src/main/java/com/nohadev/puntoventa/catalogo/dto/ProductoRequest.java
@@ -3,15 +3,17 @@ package com.nohadev.puntoventa.catalogo.dto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 public class ProductoRequest {
    private  String codigo_barras;
    private  String sku;
    private String descripcion;
-   private Double costo_promedio;
-   private Double precio_venta;
-   private Double stock_minimo;
+   private BigDecimal costo_promedio;
+   private BigDecimal precio_venta;
+   private BigDecimal stock_minimo;
    private Long categoriaId;
    private Long marcaId;
    private Boolean activo;
